requirements.txt
# Core
scikit-learn>=1.3
scikit-surprise>=1.1.3
fastapi>=0.104
uvicorn[standard]>=0.24
redis>=5.0
pydantic>=2.5
pydantic-settings>=2.1
pandas>=2.1
numpy>=1.26
pyarrow>=14.0
PyYAML>=6.0
httpx>=0.25
python-dotenv>=1.0
scipy>=1.11

# Dev
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.23
ruff>=0.1
pre-commit>=3.0
fakeredis>=2.20
